Refactor Google Drive connector
